+++ app/Badge.php
@@ -3,9 +3,14 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Spatie\MediaLibrary\HasMedia;
+use Spatie\MediaLibrary\InteractsWithMedia;
 
-class Badge extends Model
-{
+class Badge extends Model implements HasMedia
+{    
+    
+    use InteractsWithMedia;
+
     protected $fillable = [
         'title',
         'description',
@@ -17,6 +22,13 @@ class Badge extends Model
         'xp',
         'gold',
     ];
+
+    public function registerMediaCollections() : void 
+    {
+        $this
+            ->addMediaCollection('badge')
+            ->singleFile();
+    }
 
     public function classroom() {
         return $this->belongsTo(Classroom::class);