+++ app/Http/Controllers/BadgeController.php
@@ -6,6 +6,9 @@ use App\Badge;
 use App\Classroom;
 use App\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Facades\Image;
+
 
 class BadgeController extends Controller
 {
@@ -40,27 +43,27 @@ class BadgeController extends Controller
         return view('badges.create', compact('class', 'badge'));
     }
 
-    public function update($badge)
-    {
-        $badge = Badge::findOrFail($badge);
-        $class = Classroom::where('id', '=', $badge->classroom_id)->firstOrFail();
-        $this->authorize('update', $class);
-        try {
-            $badge->update($this->validateFormat(request()));
-            return [
-                "message" => __('success_error.update_success'),
-                "type" => "success",
-                "icon" => "check"
-            ];
-        } catch (\Throwable $th) {
-            return [
-                "message" => __('success_error.error'),
-                "type" => "times",
-                "type" => "error"
-            ];
-            return $th;
-        }
-    }
+    // public function update($badge)
+    // {
+    //     $badge = Badge::findOrFail($badge);
+    //     $class = Classroom::where('id', '=', $badge->classroom_id)->firstOrFail();
+    //     $this->authorize('update', $class);
+    // try {
+    //     $badge->update($this->validateFormat(request()));
+    //     return [
+    //         "message" => __('success_error.update_success'),
+    //         "type" => "success",
+    //         "icon" => "check"
+    //     ];
+    // } catch (\Throwable $th) {
+    //     return [
+    //         "message" => __('success_error.error'),
+    //         "type" => "times",
+    //         "type" => "error"
+    //     ];
+    //     return $th;
+    // }
+    // }
 
     public function toggle()
     {
@@ -97,17 +100,57 @@ class BadgeController extends Controller
         $this->authorize('update', $class);
         $data = $this->validateFormat(request());
 
-        $badge = Badge::create($data);
-        $class->badges()->save($badge);
+        if (request()->id) {
+            $badge = Badge::where('id', request()->id)->where('classroom_id', $class->id)->firstOrFail();
+            try {
+                $badge->update($data);
+            } catch (\Throwable $th) {
+                return [
+                    "message" => __('success_error.error'),
+                    "type" => "times",
+                    "type" => "error"
+                ];
+                return $th;
+            }
+        } else {
+            $badge = Badge::create($data);
+            $class->badges()->save($badge);
+        }
+
+
+        if (request()->file('image') && request()->type == 1) {
+
+            $badge->addMedia(request()->file('image'))
+                ->toMediaCollection('badge');
+
+            $itemPath = $badge->getMedia('badge')->first();
+            $imgPath = $itemPath->collection_name . "/" . $itemPath->uuid . '/' . $itemPath->file_name;
+            $path = Storage::disk('public')->path('/') . $imgPath;
+            if ($itemPath->mime_type != "image/gif" || $itemPath->size >= 500000) {
+                Image::make($path)->resize(80, 80)->save();
+            }
+            $badge->update(['image' => '/storage/' . $imgPath]);
+        }
+
+        $message = __('success_error.add_success');
+        if (request()->id)
+            $message = __('success_error.update_success');
+        return [
+            "message" => $message,
+            "type" => "check",
+            "type" => "success"
+        ];
+
         return redirect('/classroom/' . $code . '/badges');
     }
 
     public function validateFormat($request)
     {
         return $request->validate([
-            'icon' => ['required', 'string'],
+            'icon' => ['nullable', 'string'],
             'title' => ['required', 'string'],
             'description' => ['nullable', 'string'],
+            'type' => ['required', 'numeric'],
             'xp' => ['required', 'numeric'],
             'hp' => ['required', 'numeric'],
             'gold' => ['required', 'numeric'],